requirements.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
mypy==1.9.0
mypy-extensions==1.0.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-decouple==3.8
setuptools==68.2.2
sqlparse==0.5.0
typing_extensions==4.11.0
wheel==0.41.2
